fix(layer5): files 端点全新项目守卫 + read/write/delete body/query path + file_count 类型
端到端烟测发现一批 web 适配缺陷,均为"空状态掩盖"(空 team/无文件时误判通过):

files 端点全新项目 storage base 守卫(src-server/src/routes/files.rs):
- list_files/read_file 加 base.exists() 短路(全新项目返回 []/404,非 500);
  stat_file/raw_file 已有此守卫,list/read 漏了。
- upload_file/write_file 加 ensure_dir(&base)(写端点必须建目录);upload 额外
  ensure 目标子目录 + 拒绝含 .. 的 dest_subdir 防 traversal。
- delete_file 加 base.exists() 短路(404)。

read/write/delete path 来源(核心 web fs 修复):
- 早期实现用 URL *path(apiClient URL /read?path= → *path="read" 仅占位),
  body/query path 被忽略 → write 写 base/write、read 读 base/read,web fs 读写
  全错位(全新项目无文件时表现为 404 掩盖)。
- read_file 改用 query path(ReadQuery);write_file 用 body path(WriteRequest
  加 path 字段)+ ensure 父目录 + 防 ..;delete_file 用 body path(DeleteRequest)。

list_projects file_count 类型(src-server/src/routes/projects.rs):
- SELECT 0 as file_count(postgres int4)vs ProjectResponse.file_count:i64(int8)
  → 有 ≥1 project 时 sqlx 映射 500;空 team(0 行无映射)误判 200 长期隐藏。
- 改 0::bigint(5 处)。

测试(tests/integration/):
- files_list_test:list 全新项目子目录/根 → []/200(3)。
- files_fresh_project_test:read 404 / write 建目录 / upload 建目录+子目录 /
  delete 404 / read-write 往返内容 / delete body path(7)。
- projects_list_test:list 有 project → 200 + file_count / 空 team → [](2)。

+++ src-server/src/routes/projects.rs
@@ -165,7 +165,7 @@ async fn list_projects(
         let projects: Vec<ProjectResponse> = if let Some(ref cursor) = q.cursor {
             let (cursor_id, cursor_ts) = decode_cursor(cursor)?;
             sqlx::query_as::<_, ProjectResponse>(
-                "SELECT id, team_id, name, storage_path, created_by, created_at, 0 as file_count \
+                "SELECT id, team_id, name, storage_path, created_by, created_at, 0::bigint as file_count \
                  FROM projects \
                  WHERE team_id = $1 \
                    AND (created_at, id) < ($2::timestamptz, $3) \
@@ -180,7 +180,7 @@ async fn list_projects(
             .await?
         } else {
             sqlx::query_as::<_, ProjectResponse>(
-                "SELECT id, team_id, name, storage_path, created_by, created_at, 0 as file_count \
+                "SELECT id, team_id, name, storage_path, created_by, created_at, 0::bigint as file_count \
                  FROM projects \
                  WHERE team_id = $1 \
                  ORDER BY created_at DESC, id DESC \
@@ -207,7 +207,7 @@ async fn list_projects(
         let projects: Vec<ProjectResponse> = if let Some(ref cursor) = q.cursor {
             let (cursor_id, cursor_ts) = decode_cursor(cursor)?;
             sqlx::query_as::<_, ProjectResponse>(
-                "SELECT p.id, p.team_id, p.name, p.storage_path, p.created_by, p.created_at, 0 as file_count \
+                "SELECT p.id, p.team_id, p.name, p.storage_path, p.created_by, p.created_at, 0::bigint as file_count \
                  FROM projects p \
                  INNER JOIN team_members tm ON p.team_id = tm.team_id \
                  WHERE tm.user_id = $1 \
@@ -223,7 +223,7 @@ async fn list_projects(
             .await?
         } else {
             sqlx::query_as::<_, ProjectResponse>(
-                "SELECT p.id, p.team_id, p.name, p.storage_path, p.created_by, p.created_at, 0 as file_count \
+                "SELECT p.id, p.team_id, p.name, p.storage_path, p.created_by, p.created_at, 0::bigint as file_count \
                  FROM projects p \
                  INNER JOIN team_members tm ON p.team_id = tm.team_id \
                  WHERE tm.user_id = $1 \
@@ -262,7 +262,7 @@ async fn get_project(
 
     // Verify team membership via JOIN
     let project = sqlx::query_as::<_, ProjectResponse>(
-        "SELECT p.id, p.team_id, p.name, p.storage_path, p.created_by, p.created_at, 0 as file_count \
+        "SELECT p.id, p.team_id, p.name, p.storage_path, p.created_by, p.created_at, 0::bigint as file_count \
          FROM projects p \
          INNER JOIN team_members tm ON p.team_id = tm.team_id \
          WHERE p.id = $1 AND tm.user_id = $2",
